Import bash-4.4.18-1ubuntu1osrf1 to fix qemu support.
Because bash is part of the core system the normal debootstrap command
fails with upstream's latest version (#18).
To get around that we've packaged bash with -no-pie and shoehorn it into
the chroot.

This takes some dirty doing because bash's installation and
uninstallation hooks are written in bash, so if bash cannot run they
fail and the package doesn't install.
Circumventing this with a `dpkg --purge --force-all` is insufficient
because a failure in the prerm hook will still stop uninstallation.
So we first remove the prerm script. This should do no lasting damage as
we're immediately replacing bash and all it's files with a valid
package.